Time: 0 ms (100.00%), Space: 6.2 MB (9.44%) - LeetHub

+++ 171-excel-sheet-column-number/171-excel-sheet-column-number.cpp
@@ -4,7 +4,7 @@ public:
         long long ans=0;
         for(int i=columnTitle.size()-1; i>=0; i--)
         {
-            ans+=(long long)pow(26, columnTitle.size()-i-1)*(columnTitle[i]-'A'+1);
+            ans+=pow(26, columnTitle.size()-i-1)*(columnTitle[i]-'A'+1);
         }
         return ans;
     }
